refactor(api-agent): runtime commandを分割
