+++ src-tauri/src/commands/api_agents.rs
@@ -4,197 +4,23 @@ use crate::commands::atomic_write::atomic_write;
 use crate::commands::error::{CommandError, CommandResult};
 use chrono::Utc;
 use keyring::Entry;
-use serde::{Deserialize, Serialize};
-use serde_json::{json, Value};
 use std::collections::HashMap;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::sync::Arc;
 use tauri::{AppHandle, Emitter};
 use tokio::sync::Mutex;
 use uuid::Uuid;
 
+mod providers;
+pub mod types;
+
+use self::providers::{call_provider, provider_preset};
+use self::types::*;
+
 const KEYRING_SERVICE: &str = "vibe-editor";
-const SESSION_SCHEMA_VERSION: u32 = 1;
-const MAX_AUTO_DEPTH: u32 = 3;
-const MAX_AUTO_TURNS_PER_CHAIN: u32 = 6;
-const MAX_SKILL_BYTES: usize = 48 * 1024;
-const MAX_MESSAGE_BYTES: usize = 128 * 1024;
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ApiAgentConfig {
-    pub id: String,
-    pub name: String,
-    pub runtime: String,
-    pub provider_id: String,
-    pub model: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub custom_base_url: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub temperature: Option<f64>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub max_output_tokens: Option<u32>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub system_prompt: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub skill_ids: Option<Vec<String>>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub tool_mode: Option<String>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ApiAgentMessage {
-    pub id: String,
-    pub role: String,
-    pub content: String,
-    pub created_at: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub tool_name: Option<String>,
-}
-
-#[derive(Clone, Debug, Default, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ApiAgentUsage {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub input_tokens: Option<u32>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub output_tokens: Option<u32>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub total_tokens: Option<u32>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ApiAgentTurnLog {
-    pub generation_id: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub chain_id: Option<String>,
-    pub depth: u32,
-    pub turn_number: u32,
-    pub stop_reason: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub usage: Option<ApiAgentUsage>,
-    pub created_at: String,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ApiAgentSession {
-    pub schema_version: u32,
-    pub session_id: String,
-    pub agent_id: String,
-    pub provider_id: String,
-    pub model: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub title: Option<String>,
-    pub created_at: String,
-    pub updated_at: String,
-    pub messages: Vec<ApiAgentMessage>,
-    pub turn_logs: Vec<ApiAgentTurnLog>,
-    pub tool_mode: String,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ApiAgentSessionCreateRequest {
-    #[serde(default)]
-    pub session_id: Option<String>,
-    pub agent_id: String,
-    pub provider_id: String,
-    pub model: String,
-    #[serde(default)]
-    pub title: Option<String>,
-    #[serde(default)]
-    pub tool_mode: Option<String>,
-}
-
-#[derive(Clone, Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ApiAgentSkill {
-    pub id: String,
-    pub name: String,
-    pub body: String,
-}
-
-#[derive(Clone, Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ApiAgentSendRequest {
-    pub session_id: String,
-    pub card_instance_id: String,
-    pub generation_id: String,
-    pub agent: ApiAgentConfig,
-    pub message: String,
-    #[serde(default)]
-    pub system_prompt: Option<String>,
-    #[serde(default)]
-    pub skills: Option<Vec<ApiAgentSkill>>,
-    #[serde(default)]
-    pub chain_id: Option<String>,
-    #[serde(default)]
-    pub depth: Option<u32>,
-    #[serde(default)]
-    pub turn_budget: Option<u32>,
-}
-
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ApiAgentSendResult {
-    pub ok: bool,
-    pub generation_id: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub degraded_to_read_only: Option<bool>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub error: Option<String>,
-}
-
-#[derive(Clone, Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct ApiAgentStreamEvent {
-    session_id: String,
-    card_instance_id: String,
-    generation_id: String,
-    delta: String,
-}
-
-#[derive(Clone, Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct ApiAgentToolEvent {
-    session_id: String,
-    card_instance_id: String,
-    generation_id: String,
-    name: String,
-    status: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    detail: Option<String>,
-}
-
-#[derive(Clone, Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct ApiAgentDoneEvent {
-    session_id: String,
-    card_instance_id: String,
-    generation_id: String,
-    message: ApiAgentMessage,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    usage: Option<ApiAgentUsage>,
-    stop_reason: String,
-    turn_count: u32,
-}
-
-#[derive(Clone, Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct ApiAgentErrorEvent {
-    session_id: String,
-    card_instance_id: String,
-    generation_id: String,
-    message: String,
-}
 
 static SEND_LOCKS: once_cell::sync::Lazy<Mutex<HashMap<String, Arc<Mutex<()>>>>> =
     once_cell::sync::Lazy::new(|| Mutex::new(HashMap::new()));
-static HTTP_CLIENT: once_cell::sync::Lazy<reqwest::Client> =
-    once_cell::sync::Lazy::new(reqwest::Client::new);
 
 #[tauri::command]
 pub async fn api_agent_provider_set_key(provider_id: String, key: String) -> CommandResult<()> {
@@ -472,240 +298,6 @@ pub async fn api_agent_send(
     }
 }
 
-#[derive(Clone)]
-struct ProviderPreset {
-    adapter: &'static str,
-    base_url: String,
-    supports_tools: bool,
-}
-
-fn provider_preset(
-    provider_id: &str,
-    custom_base_url: Option<&str>,
-) -> CommandResult<ProviderPreset> {
-    let preset = match provider_id {
-        "openai" => ("openai-compatible", "https://api.openai.com/v1", true),
-        "openrouter" => ("openai-compatible", "https://openrouter.ai/api/v1", true),
-        "nvidia-nim" => (
-            "openai-compatible",
-            "https://integrate.api.nvidia.com/v1",
-            false,
-        ),
-        "groq" => ("openai-compatible", "https://api.groq.com/openai/v1", false),
-        "mistral" => ("openai-compatible", "https://api.mistral.ai/v1", true),
-        "together" => ("openai-compatible", "https://api.together.xyz/v1", false),
-        "cerebras" => ("openai-compatible", "https://api.cerebras.ai/v1", false),
-        "anthropic" => ("anthropic", "https://api.anthropic.com/v1", true),
-        "gemini" => (
-            "gemini",
-            "https://generativelanguage.googleapis.com/v1beta",
-            true,
-        ),
-        "custom-openai-compatible" => (
-            "openai-compatible",
-            custom_base_url.unwrap_or("").trim(),
-            false,
-        ),
-        _ => return Err(CommandError::validation("unknown providerId")),
-    };
-    if preset.1.is_empty() {
-        return Err(CommandError::validation("custom base URL is required"));
-    }
-    Ok(ProviderPreset {
-        adapter: preset.0,
-        base_url: preset.1.trim_end_matches('/').to_string(),
-        supports_tools: preset.2,
-    })
-}
-
-async fn call_provider(
-    provider: &ProviderPreset,
-    key: &str,
-    agent: &ApiAgentConfig,
-    system_prompt: &str,
-    messages: &[ApiAgentMessage],
-) -> anyhow::Result<(String, Option<ApiAgentUsage>, String)> {
-    match provider.adapter {
-        "anthropic" => call_anthropic(provider, key, agent, system_prompt, messages).await,
-        "gemini" => call_gemini(provider, key, agent, system_prompt, messages).await,
-        _ => call_openai_compatible(provider, key, agent, system_prompt, messages).await,
-    }
-}
-
-async fn call_openai_compatible(
-    provider: &ProviderPreset,
-    key: &str,
-    agent: &ApiAgentConfig,
-    system_prompt: &str,
-    messages: &[ApiAgentMessage],
-) -> anyhow::Result<(String, Option<ApiAgentUsage>, String)> {
-    let mut req_messages = Vec::new();
-    if !system_prompt.is_empty() {
-        req_messages.push(json!({ "role": "system", "content": system_prompt }));
-    }
-    for m in messages {
-        if m.role == "tool" {
-            continue;
-        }
-        req_messages.push(json!({ "role": m.role, "content": m.content }));
-    }
-    let mut body = json!({
-        "model": agent.model,
-        "messages": req_messages,
-        "stream": false
-    });
-    if let Some(t) = agent.temperature {
-        body["temperature"] = json!(t);
-    }
-    if let Some(max) = agent.max_output_tokens {
-        body["max_tokens"] = json!(max);
-    }
-    let value: Value = HTTP_CLIENT
-        .post(format!("{}/chat/completions", provider.base_url))
-        .bearer_auth(key)
-        .json(&body)
-        .send()
-        .await?
-        .error_for_status()?
-        .json()
-        .await?;
-    let content = value["choices"][0]["message"]["content"]
-        .as_str()
-        .unwrap_or("")
-        .to_string();
-    let stop = value["choices"][0]["finish_reason"]
-        .as_str()
-        .unwrap_or("stop")
-        .to_string();
-    Ok((content, usage_from_value(&value["usage"]), stop))
-}
-
-async fn call_anthropic(
-    provider: &ProviderPreset,
-    key: &str,
-    agent: &ApiAgentConfig,
-    system_prompt: &str,
-    messages: &[ApiAgentMessage],
-) -> anyhow::Result<(String, Option<ApiAgentUsage>, String)> {
-    let req_messages: Vec<Value> = messages
-        .iter()
-        .filter(|m| m.role == "user" || m.role == "assistant")
-        .map(|m| json!({ "role": m.role, "content": m.content }))
-        .collect();
-    let mut body = json!({
-        "model": agent.model,
-        "messages": req_messages,
-        "max_tokens": agent.max_output_tokens.unwrap_or(4096)
-    });
-    if !system_prompt.is_empty() {
-        body["system"] = json!(system_prompt);
-    }
-    if let Some(t) = agent.temperature {
-        body["temperature"] = json!(t);
-    }
-    let value: Value = HTTP_CLIENT
-        .post(format!("{}/messages", provider.base_url))
-        .header("x-api-key", key)
-        .header("anthropic-version", "2023-06-01")
-        .json(&body)
-        .send()
-        .await?
-        .error_for_status()?
-        .json()
-        .await?;
-    let content = value["content"]
-        .as_array()
-        .map(|parts| {
-            parts
-                .iter()
-                .filter_map(|p| p["text"].as_str())
-                .collect::<Vec<_>>()
-                .join("")
-        })
-        .unwrap_or_default();
-    let usage = Some(ApiAgentUsage {
-        input_tokens: value["usage"]["input_tokens"].as_u64().map(|n| n as u32),
-        output_tokens: value["usage"]["output_tokens"].as_u64().map(|n| n as u32),
-        total_tokens: None,
-    });
-    let stop = value["stop_reason"].as_str().unwrap_or("stop").to_string();
-    Ok((content, usage, stop))
-}
-
-async fn call_gemini(
-    provider: &ProviderPreset,
-    key: &str,
-    agent: &ApiAgentConfig,
-    system_prompt: &str,
-    messages: &[ApiAgentMessage],
-) -> anyhow::Result<(String, Option<ApiAgentUsage>, String)> {
-    let mut contents = Vec::new();
-    for m in messages {
-        let role = if m.role == "assistant" {
-            "model"
-        } else {
-            "user"
-        };
-        if m.role == "system" || m.role == "tool" {
-            continue;
-        }
-        contents.push(json!({ "role": role, "parts": [{ "text": m.content }] }));
-    }
-    let mut body = json!({ "contents": contents });
-    if !system_prompt.is_empty() {
-        body["systemInstruction"] = json!({ "parts": [{ "text": system_prompt }] });
-    }
-    let value: Value = HTTP_CLIENT
-        .post(format!(
-            "{}/models/{}:generateContent",
-            provider.base_url, agent.model
-        ))
-        .header("x-goog-api-key", key)
-        .json(&body)
-        .send()
-        .await?
-        .error_for_status()?
-        .json()
-        .await?;
-    let content = value["candidates"][0]["content"]["parts"]
-        .as_array()
-        .map(|parts| {
-            parts
-                .iter()
-                .filter_map(|p| p["text"].as_str())
-                .collect::<Vec<_>>()
-                .join("")
-        })
-        .unwrap_or_default();
-    let usage = Some(ApiAgentUsage {
-        input_tokens: value["usageMetadata"]["promptTokenCount"]
-            .as_u64()
-            .map(|n| n as u32),
-        output_tokens: value["usageMetadata"]["candidatesTokenCount"]
-            .as_u64()
-            .map(|n| n as u32),
-        total_tokens: value["usageMetadata"]["totalTokenCount"]
-            .as_u64()
-            .map(|n| n as u32),
-    });
-    let stop = value["candidates"][0]["finishReason"]
-        .as_str()
-        .unwrap_or("STOP")
-        .to_string();
-    Ok((content, usage, stop))
-}
-
-fn usage_from_value(value: &Value) -> Option<ApiAgentUsage> {
-    if !value.is_object() {
-        return None;
-    }
-    Some(ApiAgentUsage {
-        input_tokens: value["prompt_tokens"].as_u64().map(|n| n as u32),
-        output_tokens: value["completion_tokens"].as_u64().map(|n| n as u32),
-        total_tokens: value["total_tokens"].as_u64().map(|n| n as u32),
-    })
-}
-
 fn emit_delta(app: &AppHandle, req: &ApiAgentSendRequest, delta: &str) {
     let event_name = format!("api-agent:delta:{}", req.session_id);
     let _ = app.emit(
@@ -851,9 +443,4 @@ fn build_skills_context(skills: &[ApiAgentSkill]) -> String {
         remaining = remaining.saturating_sub(body.len());
     }
     out
-}
-
-#[allow(dead_code)]
-fn _assert_path_is_absolute(path: &Path) -> bool {
-    path.is_absolute()
 }
